+++ NEWS.md
@@ -1,5 +1,22 @@
 # workflows (development version)
 
+* Simon Couch is now the maintainer (#170).
+
+* `broom::augment()` now works correctly in the edge case where you had supplied
+  a hardhat blueprint with `composition` set to either `"matrix"` or
+  `"dgCMatrix"` (#148).
+
+* `butcher::axe_fitted()` now axes the recipe preprocessor that is stored inside
+  a workflow, which will reduce the size of the `template` data frame that is
+  stored in the recipe (#147).
+
+* `add_model()` now errors informatively if you try to add a model specification
+  that contains an unknown mode (#160).
+
+* `add_formula()` no longer silently ignores offsets supplied with `offset()`.
+  Instead, it now errors at `fit()` time with a message that encourages you to
+  use a model formula through `add_model(formula = )` instead (#162).
+
 # workflows 1.0.0
 
 * New `add_case_weights()`, `update_case_weights()`, and `remove_case_weights()`